Address review feedback on shell log tests
